SERVER-65503 Migrate FLE2 tests to core passthrough suites
GitOrigin-RevId: 53196e18e01114f59037ee9c73e1a2935a55fb19

+++ src/mongo/db/modules/enterprise/jstests/fle2/basic_delete.js
@@ -2,7 +2,8 @@
  * Test encrypted delete works
  *
  * @tags: [
- * requires_fcv_60
+ * requires_fcv_60,
+ * requires_non_retryable_writes,
  * ]
  */
 load("jstests/fle2/libs/encrypted_client_util.js");
@@ -43,12 +44,14 @@ client.assertEncryptedCollectionCounts("basic", 1, 2, 1, 3);
 res = assert.commandWorked(edb.basic.deleteOne({"last": "non-existent"}));
 assert.eq(res.deletedCount, 0);
 
-// Delete a document by case-insensitive collation
-res = assert.commandWorked(
-    edb.basic.deleteOne({"last": "marco"}, {collation: {locale: 'en_US', strength: 2}}));
-assert.eq(res.deletedCount, 1);
+if (!client.useImplicitSharding) {
+    // Delete a document by case-insensitive collation
+    res = assert.commandWorked(
+        edb.basic.deleteOne({"last": "marco"}, {collation: {locale: 'en_US', strength: 2}}));
+    assert.eq(res.deletedCount, 1);
 
-client.assertEncryptedCollectionCounts("basic", 0, 2, 2, 4);
+    client.assertEncryptedCollectionCounts("basic", 0, 2, 2, 4);
+}
 
 // Negative: Test bulk delete. Query analysis is throwing this error in the shell
 res = assert.commandFailedWithCode(dbTest.basic.runCommand({
@@ -84,34 +87,39 @@ assert.throwsWithCode(() => {
 }, 6382800);
 
 // Tests that will delete documents based on encrypted fields.
-const collName = "basic_query";
-assert.commandWorked(client.createEncryptionCollection(collName, {
-    encryptedFields:
-        {"fields": [{"path": "first", "bsonType": "string", "queries": {"queryType": "equality"}}]}
-}));
-const coll = edb[collName];
-assert.commandWorked(coll.insert({"first": "mark", "last": "marco"}));
-assert.commandWorked(coll.insert({"first": "Mark", "last": "Marco"}));
-client.assertEncryptedCollectionCounts(collName, 2, 2, 0, 2);
+if (!client.useImplicitSharding) {
+    const collName = "basic_query";
+    assert.commandWorked(client.createEncryptionCollection(collName, {
+        encryptedFields: {
+            "fields":
+                [{"path": "first", "bsonType": "string", "queries": {"queryType": "equality"}}]
+        }
+    }));
+    const coll = edb[collName];
+    assert.commandWorked(coll.insert({"first": "mark", "last": "marco"}));
+    assert.commandWorked(coll.insert({"first": "Mark", "last": "Marco"}));
+    client.assertEncryptedCollectionCounts(collName, 2, 2, 0, 2);
 
-// Delete a document by an encrypted field.
-res = assert.commandWorked(coll.deleteOne({"first": "mark"}));
-assert.eq(res.deletedCount, 1);
-client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
+    // Delete a document by an encrypted field.
+    res = assert.commandWorked(coll.deleteOne({"first": "mark"}));
+    assert.eq(res.deletedCount, 1);
+    client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
 
-// Try deleting a non-existent document.
-res = assert.commandWorked(coll.deleteOne({"first": "dev"}));
-assert.eq(res.deletedCount, 0);
-client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
+    // Try deleting a non-existent document.
+    res = assert.commandWorked(coll.deleteOne({"first": "dev"}));
+    assert.eq(res.deletedCount, 0);
+    client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
 
-// Try deleting a non-existent combination of encrypted and non-encrypted fields, with a
-// case-insensitive collation.
-res = assert.commandWorked(coll.deleteOne({$and: [{"first": "Mark"}, {"last": "non-existent"}]}));
-assert.eq(res.deletedCount, 0);
-client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
+    // Try deleting a non-existent combination of encrypted and non-encrypted fields, with a
+    // case-insensitive collation.
+    res =
+        assert.commandWorked(coll.deleteOne({$and: [{"first": "Mark"}, {"last": "non-existent"}]}));
+    assert.eq(res.deletedCount, 0);
+    client.assertEncryptedCollectionCounts(collName, 1, 2, 1, 3);
 
-// Delete with a combination of encrypted and non-encrypted fields.
-res = assert.commandWorked(coll.deleteOne({$and: [{"first": "Mark"}, {"last": "Marco"}]}));
-assert.eq(res.deletedCount, 1);
-client.assertEncryptedCollectionCounts(collName, 0, 2, 2, 4);
+    // Delete with a combination of encrypted and non-encrypted fields.
+    res = assert.commandWorked(coll.deleteOne({$and: [{"first": "Mark"}, {"last": "Marco"}]}));
+    assert.eq(res.deletedCount, 1);
+    client.assertEncryptedCollectionCounts(collName, 0, 2, 2, 4);
+}
 }());
